FIX(activity): align recognized-hour aggregation
- Preserve manually granted activity exemptions during member refreshes.

- Match admin totals to My Page session and project calculations.

+++ src/main/java/org/poolc/api/member/dto/MemberResponseWithHour.java
@@ -4,21 +4,23 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import lombok.Getter;
 import org.poolc.api.member.domain.Member;
 
+import java.math.BigDecimal;
+
 @Getter
 public class MemberResponseWithHour {
 
     private final MemberResponse member;
-    private final Long hour;
+    private final BigDecimal hour;
     private final Boolean isExcepted;
-    
+
     @JsonCreator
-    public MemberResponseWithHour(MemberResponse member, Long hour, Boolean isExcepted) {
+    public MemberResponseWithHour(MemberResponse member, BigDecimal hour, Boolean isExcepted) {
         this.member = member;
         this.hour = hour;
         this.isExcepted = isExcepted;
     }
 
-    public static MemberResponseWithHour of(Member member, Long hour) {
+    public static MemberResponseWithHour of(Member member, BigDecimal hour) {
         return new MemberResponseWithHour(MemberResponse.of(member), hour, member.getIsExcepted());
     }
 }
